Improve delta tests + fix delta+encrypt bug

+++ include/delta.h
@@ -67,7 +67,7 @@ int wb_diff_init(WB_DIFF_CTX *ctx, uint8_t *src_a, uint32_t len_a, uint8_t *src_
 int wb_diff(WB_DIFF_CTX *ctx, uint8_t *patch, uint32_t len);
 int wb_patch_init(WB_PATCH_CTX *bm, uint8_t *src, uint32_t ssz, uint8_t *patch, uint32_t psz);
 int wb_patch(WB_PATCH_CTX *ctx, uint8_t *dst, uint32_t len);
-int wolfBoot_get_delta_info(uint8_t part, int inverse, uint32_t **img_offset, uint16_t **img_size);
+int wolfBoot_get_delta_info(uint8_t part, int inverse, uint32_t **img_offset, uint32_t **img_size);
 
 #endif
